requirements.txt
django==6.0.3
djangorestframework==3.17.0
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
bcrypt==5.0.0
drf-spectacular==0.29.0
